[ refactor ] Factoring out the unsafe parts
It is currently impossible to typecheck a development relying on
the standard library with the --safe option because it does not
isolate properly its uses of unsafe features

The main culprits are "primTrustMe" and "erase" but also
"NON_TERMINATING" or "postulate" + "COMPILED" in the IO module.

This patch isolates the unsafe-tainted functions in Unsafe.*
modules (each of which re-exports its safe counterpart so that
switching from one to the other is as simple as adding the
prefix "Unsafe." to the module's name in the import statement).