[ update ] container as a basis for W and M
Until now W and M were based on a pair of a set for shapes and one
for positions. The only distinction with containers at the time were
that shapes and positions could live at different levels.

Now that the containers are level polymorphic (PR #346), we can use
them as the basis for M and W.