Relax decidability module parameter of Algebra.Solver.Ring
This replaces the `Decidable` module parameter in `Algebra.Solver.Ring`
by a new predicate `WeaklyDecidable`. It only provides a subset of those
equality proofs that `Decidable` can provide, and does not provide
inequality proofs.

As a result, ring solvers can now be used in more general situations in
which decidable equality is not available. For this, a new module
`Algebra.Solver.Ring.NaturalCoefficients.Default` was added which
instantiates the natural coefficients solver for any
`CommutativeSemiring`, such as `×-⊎-commutativeSemiring` in
`Function.Related.TypeIsomorphisms.Solver`.