[ refactor ] Cleaning up Relation.Nullary.Decidable via Core
In order to define generic projections for n-ary products we need
to be able to import `Data.Fin.Base`. It itself currently imports
`Relation.Nullary.Decidable` which pulls all sorts of `Function.X`
depending on `Relation.Binary.PropositionalEquality` and thus on
n-ary product (because of generic congₙ and substₙ).

This commit cleans up the dependencies of the modules involved in
the cycle and splits off `Relation.Nullary.Decidable.Core` for
simple definitions.