[ new ] Data.Vec.Bounded for bounded vectors
I don't find Data.BoundedVec easy to use. Additionally this module
is fully lazy in the (irrelevant) proofs that the current length
is smaller than the bound.

Had to introduce `Data.Vec.Base` and `Data.Vec.Bounded.Base` to break
the dependency cycle between `Data.Vec` and `Data.Vec.Bounded` because
of the definitions of `filter`, `takeWhile`, and `dropWhile`.