[ fix #826 ] Coercible type class
To cross FFI boundaries easily, we can use `coerce` to convert
between the FFI-friendly types and the ones actually use in the
core library with all the functions we know and love.

Note that I had to swap Data.Maybe.Base's constructors so that
they would appear in the same order as the ones for Maybe as
defined in ghc's base.