[ refined ] bugfixes
bugfixes are now stored in a directory.

The file named "others" contains basic items: each paragraph is
a bugfix and all of these get grouped in a `#### Others` section
across PRs.

The others files are meant to host a section each. They can use
markdown and are embedded as is.