[ agda/agda#4560 ] 'eta-equality' or 'pattern' needed now
The fix of agda/agda#4560 in PR agda/agda#4611 requires the record
directive 'pattern' if you want to pattern match on a record constructor
that does not have eta-equality.  Eta-equality is automatically turned
off for some recursive records, such as the present one.
Here, we turn eta on to be compatible with Agda without the PR.

Alternatively, we could add 'pattern' instead of 'eta-equality'.