Remove single backslash in string literal, and update changelog.
Using a single backslash is interpreted as an escape sequence. Therefore \prime inside a string literal is an I know we can add two, but we may as well not use any here.  This is supposed to be a fix for : Lexical error in string or character literal: bad escape code.