agda-stdlib-utils: drop GHC < 8.6, modernise .cabal file, update CI
With Agda-2.6.4 dropping GHC < 8.6, we can follow suit here.
This encourages us to also bump the `cabal-version`,
which allows us to modernise the `.cabal` file.

Update the Haskell-CI to the latest versions.
Also insert some `paths` triggers via the patching mechanism,
so that this workflow does not run on the usual PRs that
just contribute to the Agda code in the library.