Switch to Agda 2.8.0-rc3 by merging `experimental` into `master` (#2755)
* agda/agda#7674 eta-expand fields in record expression

Agda PR #7674 requires some implicit arguments explicitly in some record
expression given since variance info got refined (invariant to
non-variant) and hence solutions are no longer unique.

* Workaround for Agda 2.7.0 (and lower, I suppose)

* Correct option --warn to --warning for agda/agda#7966

The official name of the option is `--warning` and the prefix `--warn` is only
accepted because we have no options like `--warner` or
`--warner-brothers-owns-us`, but we might have so in the future.

* Update README and CI for v2.8.0-rc3

---------

Co-authored-by: Andreas Abel <andreas.abel@ifi.lmu.de>